msmarco eval

+++ FlagEmbedding/evaluation/msmarco/utils/data_loader.py
@@ -9,6 +9,9 @@ import gzip
 import shutil
 import tarfile
 import json
+import csv
+import sys
+csv.field_size_limit(sys.maxsize)
 
 from tqdm import tqdm, trange
 from huggingface_hub import snapshot_download
@@ -28,9 +31,10 @@ def download_file(
         source_file_path = os.path.join(directory, url.split('/')[-1])
         file_path = os.path.join(directory, filename)
 
-        with open(source_file_path, 'wb') as file:
-            for chunk in tqdm(response.iter_content(chunk_size=8192), desc="Downloading"):
-                file.write(chunk)
+        if not os.path.exists(source_file_path):
+            with open(source_file_path, 'wb') as file:
+                for chunk in tqdm(response.iter_content(chunk_size=8192), desc="Downloading"):
+                    file.write(chunk)
 
         logger.warning(f"File downloaded: {source_file_path}")
 
@@ -67,40 +71,34 @@ def download_file(
                         results[str(list(tmp.keys())[0])] = list(tmp.values())[0]
             else:
                 results = {}
-                with open(temp_file_path) as f:
-                    for line in f:
-                        if '\t' in line:
-                            tmp = line.strip().split('\t')
-                        else:
-                            tmp = line.strip().split()
-                        if len(tmp) == 2:
-                            results[str(tmp[0])] = tmp[1]
-                        else:
-                            if 'http' not in str(tmp[1]):
-                                results[str(tmp[0])] = (tmp[2] + ' ' + tmp[3]).strip()
+                with open(temp_file_path, newline='', encoding='utf-8') as f:
+                    reader = csv.reader(f, delimiter='\t')
+                    for row in reader:
+                        if len(row) == 2:
+                            results[row[0]] = row[1]
+                        elif len(row) > 3:
+                            if 'http' in row[1]:
+                                results[row[0]] = f"{row[2]} {row[3]}".strip()
                             else:
-                                if str(tmp[0]) not in results.keys():
-                                    results[str(tmp[0])] = {}
-                                results[str(tmp[0])][str(tmp[2])] = int(tmp[3])
+                                if row[0] not in results:
+                                    results[row[0]] = {}
+                                results[row[0]][row[2]] = int(row[3])
             os.remove(temp_file_path)
 
         else:
             results = {}
-            with open(source_file_path) as f:
-                for line in f:
-                    if '\t' in line:
-                        tmp = line.strip().split('\t')
-                    else:
-                        tmp = line.strip().split()
-                    if len(tmp) == 2:
-                        results[str(tmp[0])] = tmp[1]
-                    else:
-                        if 'http' not in str(tmp[1]):
-                            results[str(tmp[0])] = (tmp[2] + ' ' + tmp[3]).strip()
+            with open(temp_file_path, newline='', encoding='utf-8') as f:
+                reader = csv.reader(f, delimiter='\t')
+                for row in reader:
+                    if len(row) == 2:
+                        results[row[0]] = row[1]
+                    elif len(row) > 3:
+                        if 'http' in row[1]:
+                            results[row[0]] = f"{row[2]} {row[3]}".strip()
                         else:
-                            if str(tmp[0]) not in results.keys():
-                                results[str(tmp[0])] = {}
-                            results[str(tmp[0])][str(tmp[2])] = int(tmp[3])
+                            if row[0] not in results:
+                                results[row[0]] = {}
+                            results[row[0]][row[2]] = int(row[3])
             os.remove(source_file_path)
 
         with open(file_path, 'w') as f:
@@ -211,12 +209,12 @@ class MSMARCODataLoader(AbsDataLoader):
                             'qrels-dl20.json'
                         )
             else:
-                if not os.path.exists(corpus_path):
-                    corpus_path = download_file(
-                        "https://msmarco.z22.web.core.windows.net/msmarcoranking/msmarco-docs.tsv.gz",
-                        self.dataset_dir,
-                        'corpus.json'
-                    )
+                # if not os.path.exists(corpus_path):
+                #     corpus_path = download_file(
+                #         "https://msmarco.z22.web.core.windows.net/msmarcoranking/msmarco-docs.tsv.gz",
+                #         self.dataset_dir,
+                #         'corpus.json'
+                #     )
                 if split == 'dev':
                     if not os.path.exists(queries_path):
                         queries_path = download_file(
@@ -280,5 +278,5 @@ class MSMARCODataLoader(AbsDataLoader):
         new_queries = {}
         for k in queries.keys():
             if k in rels.keys():
-                new_queries[k] = rels[k]
+                new_queries[k] = queries[k]
         return datasets.DatasetDict(new_queries)
